+++ examples/load_windfield.jl
@@ -4,7 +4,8 @@ set_data_path("data")
 set = load_settings("system.yaml"; relax=true)
 am::AtmosphericModel = AtmosphericModel(set)
 
-x,y,z,u,v,w,param = AtmosphericModels.load_windfield(am, 5.324)
+u, v, w, param, v_wind_gnd = AtmosphericModels.load_windfield(am, 5.324)
 
-@info "Windfield dimensions: $(size(x)), $(size(y)), $(size(z))"
-@info "Parameters: $param"
+@info "Windfield dimensions: $(size(u))"
+@info "Grid axes: $(AtmosphericModels.grid_axes(am))"
+@info "Parameters: $param, generated for $v_wind_gnd m/s"